modify: file divided for VsyncProvider
