+++ lib/src/in_app_notification.dart
@@ -5,6 +5,8 @@ import 'package:flutter/material.dart';
 import 'package:in_app_notification/src/interact_animation_controller.dart';
 import 'package:in_app_notification/src/size_listenable_container.dart';
 
+part 'vsync_provider.dart';
+
 @visibleForTesting
 const notificationShowingDuration = Duration(milliseconds: 350);
 
@@ -82,57 +84,6 @@ class InAppNotification extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return _VsyncProvider(child: child);
-  }
-}
-
-class _VsyncProvider extends StatefulWidget {
-  const _VsyncProvider({
-    Key? key,
-    required this.child,
-  }) : super(key: key);
-
-  final Widget child;
-
-  @override
-  State<_VsyncProvider> createState() => __VsyncProviderState();
-}
-
-class __VsyncProviderState extends State<_VsyncProvider>
-    with TickerProviderStateMixin {
-  late final _showController = AnimationController(
-    vsync: this,
-    duration: notificationShowingDuration,
-  );
-
-  late final _verticalAnimationController = VerticalInteractAnimationController(
-    vsync: this,
-    duration: notificationShowingDuration,
-  );
-
-  late final _horizontalAnimationController =
-      HorizontalInteractAnimationController(
-    vsync: this,
-    duration: notificationHorizontalAnimationDuration,
-  );
-
-  @override
-  Widget build(BuildContext context) {
-    return _NotificationController(
-      state: _NotificationState(
-        showController: _showController,
-        verticalAnimationController: _verticalAnimationController,
-        horizontalAnimationController: _horizontalAnimationController,
-      ),
-      child: widget.child,
-    );
-  }
-
-  @override
-  void dispose() {
-    _showController.dispose();
-    _verticalAnimationController.dispose();
-    _horizontalAnimationController.dispose();
-    super.dispose();
   }
 }
